requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
gcloud==0.18.3
googleapis-common-protos==1.51.0
gpsphoto==2.2.3
httplib2==0.17.0
idna==2.8
jws==0.1.3
oauth2client==4.1.2
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.4
Pyrebase4==4.3.0
python-jwt==2.0.1
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
